pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azx"
version = "0.0.1"
description = "azx"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "jax",
  "chex",
  "mctx",
  "orbax",
  "dm-haiku",
  "optax",
  "jumanji",
  "flashbax",
  "rlax"
]

[tool.setuptools.packages.find]
include = ["azx*"]

